requirements.txt
certifi==2018.8.13
chardet==3.0.4
dj-database-url==0.4.2
Django==2.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
mysqlclient==1.3.13
psycopg2==2.7.4
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
Unipath==1.1
urllib3==1.23
whitenoise==4.0
